Pass explicit list of file extensions
Rome chokes if you ask it to format a non-javascript file (like yaml),
so relying on `types: [text]` will cause problems. Switching to an
explicit list of extensions isn't great because it limits rome for some
folks who may use exotic extensions for their javascript files. We'll
cross that bridge we come to it.